Use union for SMF (#12)
It's much nicer if there's just one output format for standard MIDI files, so unify the formats into one using a union. As it happens, I'm exposing exactly the same structure to the user for Format 0 and Format 1, so a union isn't an obvious choice here. However, when support for Format 2 is added, it will make sense.