SmfPlugin now uses Type/Values (#31)
The SMF plugin has been converted to use the type system rather than the features system. In particular, Source and Target formats now operate on the same type of data.

This means you can directly wire any part of the hierarchy in a source SMF file to its equivalent in a target SMF file, include the file contents as a whole. Parts of the copy in the target parts can then be overwritten (subject to limits on structural modification).

* SMF0 uses recordOfMidiTracks

* MidiTrackAndChannel gets optional tracks.

* Better identifiers for extra channels

* smfTargetModel -> smfFeature

* SmfSequenceType -> SmfSequence

* smfSequenceType file -> smfSequence

* Use new UUIDs

* SmfParser::Format

* Tweak

* Tidy

* Now parse to SmfFeature

* Adapt saveLoadTests.

* Adapt percussionTests.cpp

* Adapt testSuiteTests

* Remove source model

* Save extra channels

* Update sample